# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Posts
...
  /chat/conversations:
    get:
      operationId: getConversations
      tags: [chat]
      summary: Get user conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationsResponse"

    post:
      operationId: createConversation
      tags: [chat]
      summary: Create or get a direct conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chat/chatrooms:
    get:
      operationId: getChatrooms
      tags: [chat]
      summary: Get public chatrooms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chatrooms list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatroomsResponse"
  /chat/chatrooms/{chatroomId}/messages:
    get:
      operationId: getChatroomMessages
      tags: [chat]
      summary: Get chatroom messages
      security:
        - bearerAuth: []
      parameters:
        - name: chatroomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"

    post:
      operationId: sendChatroomMessage
      tags: [chat]
      summary: Send a chatroom message
      security:
        - bearerAuth: []
      parameters:
        - name: chatroomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # Wallet
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth schemas
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, provider]
          default: student
        college:
          type: string
        program:
          type: string
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User schemas
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        college:
          type: string
          nullable: true
        program:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, provider, admin]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - followersCount
        - followingCount
        - postsCount
        - createdAt
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        unreadCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - participants
        - unreadCount
        - updatedAt
    ConversationsResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
      required:
        - conversations
    CreateConversationRequest:
      type: object
      properties:
        participantId:
          type: string
      required:
        - participantId
    Message:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderAvatar:
          type: string
          nullable: true
        conversationId:
          type: string
          nullable: true
        chatroomId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - senderId
        - senderName
        - createdAt
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: string
          nullable: true
      required:
        - messages
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Chatroom:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        memberCount:
          type: integer
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - memberCount
        - updatedAt
    ChatroomsResponse:
      type: object
      properties:
        chatrooms:
          type: array
          items:
            $ref: "#/components/schemas/Chatroom"
      required:
        - chatrooms

    # Wallet schemas
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [like, comment, follow, message, service, wallet]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        data:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - body
        - isRead
        - createdAt
    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required:
        - notifications
        - unreadCount