# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string | null;
  bio?: string | null;
  college?: string | null;
  program?: string | null;
  role: UserRole;
  followersCount: number;
  followingCount: number;
  postsCount: number;
  createdAt: Date;
}